pyproject.toml
[project]
name = "veimagex-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
    "volcengine>=1.0.177",
    "starlette>=0.14.2",
    "typing-extensions>=4.0.0",
    "pydantic>=1.8.2",
    "retry>=0.9.2"
]
